files Updated package name and manually tested the FileHandler.
