--- CourseManagementSystem/.idea/libraries/Maven__org_junit_jupiter_junit_jupiter_api_5_8_1.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.junit.jupiter:junit-jupiter-api:5.8.1">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/junit/jupiter/junit-jupiter-api/5.8.1/junit-jupiter-api-5.8.1.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/junit/jupiter/junit-jupiter-api/5.8.1/junit-jupiter-api-5.8.1-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/junit/jupiter/junit-jupiter-api/5.8.1/junit-jupiter-api-5.8.1-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
