--- CourseManagementSystem/.idea/libraries/Maven__org_junit_platform_junit_platform_commons_1_8_1.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.junit.platform:junit-platform-commons:1.8.1">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/junit/platform/junit-platform-commons/1.8.1/junit-platform-commons-1.8.1.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/junit/platform/junit-platform-commons/1.8.1/junit-platform-commons-1.8.1-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/junit/platform/junit-platform-commons/1.8.1/junit-platform-commons-1.8.1-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
