+++ CourseManagementSystem/.idea/libraries/Maven__org_testng_testng_7_5.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.testng:testng:7.5">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/testng/testng/7.5/testng-7.5.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/testng/testng/7.5/testng-7.5-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/testng/testng/7.5/testng-7.5-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
