+++ CourseManagementSystem/src/main/java/com/company/cms/CMSController.java
@@ -1,0 +1,14 @@
+package com.company.cms;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.Label;
+
+public class CMSController {
+    @FXML
+    private Label welcomeText;
+
+    @FXML
+    protected void onHelloButtonClick() {
+        welcomeText.setText("Welcome to JavaFX Application!");
+    }
+}
