+++ CourseManagementSystem/src/main/java/com/company/cms/FileHandling/Filename.java
@@ -1,0 +1,11 @@
+package com.company.cms.FileHandling;
+
+public enum Filename {
+    ASSIGNMENTS,
+    COURSEMODULES,
+    COURSES,
+    RESULTS,
+    COURSEADMINISTRATORS,
+    INSTRUCTORS,
+    STUDENTS
+}
