--- CourseManagementSystem/src/main/java/com/example/cms/FileHandling/Filename.java
@@ -1,11 +1,0 @@
-package com.example.cms.FileHandling;
-
-public enum Filename {
-    ASSIGNMENTS,
-    COURSEMODULES,
-    COURSES,
-    RESULTS,
-    COURSEADMINISTRATORS,
-    INSTRUCTORS,
-    STUDENTS
-}
