--- CourseManagementSystem/src/main/resources/com/example/cms/create-account-view.fxml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.VBox?>
-
-<?import javafx.scene.control.Button?>
-<VBox alignment="CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="com.example.cms.CMSController">
-    <padding>
-        <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
-    </padding>
-
-    <Label fx:id="createAccountHeader">Create Account</Label>
-    <Label fx:id="createAccountUsername">Username:</Label>
-    <Label fx:id="createAccountPassword">Password:</Label>
-</VBox>
