--- CourseManagementSystem/src/main/resources/com/example/cms/start-view.fxml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.VBox?>
-
-<?import javafx.scene.control.Button?>
-<VBox alignment="CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="com.example.cms.CMSController">
-    <padding>
-        <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
-    </padding>
-
-    <Label fx:id="startHeader">Course Management System</Label>
-    <Button fx:id="createAccountButton">Create Account</Button>
-    <Button fx:id="loginButton">Login</Button>
-</VBox>
